Choose color for alpha channel replacement
Added ability to change the color that is used to replace the alpha channel with.
This can be done either by adding the background HTML parameter, or by setting a default/fallback color for a specific format type in the plugin options in vite.config.
Both can be a string starting with or without #, and followed by 2, 3 or 6 hexadecimal digits (0-9,a-f). The latter can also be a RGBA object as used in sharp itself (although I left it undocumented in the readme to avoid confusion since most people use hex anyways).
I also removed some superfluous async/awaits.
Some choices had to be made when making this, like the choice of using the OptimizationOptions for passing in a default/fallback value for the background color. This may not be correct regarding it not being a OptimizationOption, but it did seem a user-friendly and logic place to place it (the OptimizationOptions typing looked like it wasn't working anyways tbh). Anyways, if this choice or other choices (changing handleFormat to a handleConversion and the refactoring of its parameters, etc) are not agreed upon, let me know so I can change it.