devMode lets you choose behaviour for vite --host
With devMode option you can choose to not process all images this plugin normally would, saving time when focussing on other parts of your project.
When the focus is on the images, normal behaviour can be used by setting devMode to "transform", which is in fact the default.
Setting devMode to "skip" makes it skip all image tranformations, "formatOnly" makes sure it is in the correct image format, but won't apply other options.